backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.2.0
scipy==1.12.0
torch==2.2.0
torchvision==0.17.0
pydantic==2.6.0
